+++ src/primitive/defs.rs
@@ -1098,6 +1098,26 @@ primitive!(
     /// ex: ⌕ 1_2 . ↯4_4⇡3
     /// ex: ⌕ [1_2 2_0] . ↯4_4⇡3
     (2, Find, DyadicArray, ("find", '⌕')),
+    /// Mask the occurences of one array in another
+    ///
+    /// Occurences of the first array in the second array will be marked with increasing numbers.
+    /// While [find] only marks the start of each occurence, [mask] marks the entire occurence.
+    /// ex: # Experimental!
+    ///   : ⍝ "ab" "abracadabra"
+    /// ex: # Experimental!
+    ///   : ⍝ [1 2 3].[0 1 2 3 1 2 3 4 5 1 2 3 4 5 6]
+    /// Increasing numbers are used so that adjacent occurences can be distinguished.
+    /// An occurence that would overlap with a previous occurence is not marked.
+    /// ex: # Experimental!
+    ///   : ⍝ [3 4 3 4].[0 3 4 3 4 1 2 0 0 3 4 3 4 0]
+    ///
+    /// Arbtrary rank arrays are supported.
+    /// The needle's rank must be `less or equal` the haystack's rank.
+    /// ex: # Experimental!
+    ///   : ⍝,, 3_4 ↯2_3⇡6
+    /// ex: # Experimental!
+    ///   : ⍝,, [1_2 5_6] [1_2_3_4 5_6_1_2 7_8_5_6 4_3_1_2]
+    (2, Mask, DyadicArray, ("mask", '⍝')),
     /// Check if each row of one array exists in another
     ///
     /// ex: ∊ 2 [1 2 3]